feat(components/verifyemail/): added AlreadySignedInPrompt.tsx option to fixed verify email page
verify email page now has three states: isAlreadySignedIn, isFailed, and defualt (loading). added
the isAlreadySignedIn option due to cannot use guard for verify email route (interferes with
Stack-Auth native auto-redir). this covers the case where somehow a signed-in user still wishes to
access the verify-email route. if so, the user is prompted to either go to dashboard or to sign out,
currently just link to sign-out page instead of immediately to modularise duties